simd: correct suggestVectorLength
RISC-V vectors are VLA, or vector length agnostic. This means that the compiler can set the amount of lanes on a register as it sees fit. The limitation is in the element width per lane, which is 64 bits max. Returning 64 bits here makes more sense because working with 64 bit element vectors is easier to optimize and provides cleaner codegen.